+++ blocks/list/src/components/ListBlock/ListBlock.tsx
@@ -1,6 +1,6 @@
 import { BlockProps } from '@appsemble/react';
 import { Loader } from '@appsemble/react-components';
-import { remapData } from '@appsemble/utils/remap';
+import { remapData } from '@appsemble/utils';
 import React from 'react';
 import { FormattedMessage, InjectedIntlProps } from 'react-intl';