+++ packages/react-components/components/Location/Location.tsx
@@ -1,6 +1,7 @@
 import 'leaflet/dist/leaflet.css';
 
 import { BlockProps } from '@appsemble/react';
+import { Theme } from '@appsemble/types';
 import {
   CircleMarker,
   Icon,
@@ -21,6 +22,7 @@ export interface LocationProps {
   latitude: number;
   longitude: number;
   mapOptions: MapOptions;
+  theme: Theme;
 }
 
 /**
@@ -37,11 +39,11 @@ export default class Location extends React.Component<LocationProps & BlockProps
       latitude,
       longitude,
       mapOptions,
-      reactRoot,
+      theme: { primaryColor, tileLayer },
     } = this.props;
 
     const locationMarker = new CircleMarker(null, {
-      color: getComputedStyle(reactRoot).getPropertyValue('--primary-color'),
+      color: primaryColor,
     });
 
     const map = new Map(this.ref.current, {
@@ -49,7 +51,7 @@ export default class Location extends React.Component<LocationProps & BlockProps
       zoom: 16,
       center: [latitude, longitude],
       layers: [
-        new TileLayer('https://cartodb-basemaps-c.global.ssl.fastly.net/light_all/{z}/{x}/{y}.png'),
+        new TileLayer(tileLayer),
         new Marker([latitude, longitude], {
           icon: new Icon({
             iconUrl,