+++ packages/react/index.test.tsx
@@ -36,6 +36,17 @@ beforeEach(() => {
       addCleanup() {},
       showMessage() {},
     },
+    theme: {
+      primaryColor: '#ff0000',
+      linkColor: '#ff7f00',
+      successColor: '#ffff00',
+      infoColor: '#00ff00',
+      warningColor: '#0000ff',
+      dangerColor: '#4b0082',
+      splashColor: '#8b00ff',
+      themeColor: '#ffffff',
+      tileLayer: 'https://cartodb-basemaps-c.global.ssl.fastly.net/light_all/{z}/{x}/{y}.png',
+    },
   };
 });